main.py
import os
import logging
from dotenv import load_dotenv
import streamlit as st

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)

def main():
    st.title("PitGenius: F1 Race Strategy Optimizer")
    
    # Sidebar for input parameters
    st.sidebar.header("Race Parameters")
    
    # Race selection
    race_weekend = st.sidebar.selectbox(
        "Select Race Weekend",
        ["2024 Bahrain GP", "2024 Saudi Arabian GP", "2024 Australian GP"]
    )
    
    # Weather conditions
    weather = st.sidebar.selectbox(
        "Weather Conditions",
        ["Dry", "Wet", "Mixed"]
    )
    
    # Main content area
    st.header(f"Strategy Analysis for {race_weekend}")
    
    # Placeholder for strategy recommendations
    st.subheader("Recommended Strategy")
    st.info("Strategy recommendations will appear here")
    
    # Placeholder for visualization
    st.subheader("Race Simulation")
    st.write("Race simulation visualization will appear here")

if __name__ == "__main__":
    try:
        main()
    except Exception as e:
        logger.error(f"An error occurred: {str(e)}")
        st.error("An error occurred. Please check the logs for details.")
