requirements.txt
pandas>=1.5.0
numpy>=1.20.0
matplotlib>=3.5.0
seaborn>=0.12.0
fastf1>=3.0.0
requests>=2.28.0
python-dotenv>=0.19.0
